* add tqdm for model iteration progress bar * dont initialize_netcdf twice for Model or Process/ConservativeProcess * add Process class name to netcdf output

+++ pywatershed/base/control.py
@@ -198,8 +198,8 @@ class Control(Accessor):
         self._end_time = new_end_time
         assert self._end_time - self._start_time > 0
         self._n_times = (
-            (self._end_time - self._start_time) / self._time_step
-        ) + 1
+            int((self._end_time - self._start_time) / self._time_step) + 1
+        )
         return
 
     def edit_n_time_steps(self, new_n_time_steps: int):
